changing ints to doubles

+++ ddd/src/aula2/exercicioScanner/ExercicioScanner.java
@@ -7,13 +7,13 @@ public class ExercicioScanner {
 		Scanner input = new Scanner(System.in);
 		
 		System.out.println("Digite a sua primeiro: ");
-		int nota1 = input.nextInt();
+		double nota1 = input.nextInt();
 		
 		System.out.println("Digite a sua segunda nota:");
-		int nota2 = input.nextInt();
+		double nota2 = input.nextInt();
 		
 		System.out.println("Digite a sua terceira nota:");
-		int nota3 = input.nextInt();
+		double nota3 = input.nextInt();
 		input.close();
 		
 		double media = (nota1 + nota2 + nota3) / 3;
